KioskP Backup Copy Post
This was the copy found on backup drives, though the source of the
original content may have been changed from the location I looked in.
That said, I seem to recall that there was one version of KIosk that
didn't necessarily go anywhere except color changing, and this might
have been that version.